restructured unit tests
